Revert "Added Sprints, Sprint states. Added complementary tests and updated old tests"
