+++ .gitignore
@@ -6,4 +6,3 @@ riderModule.iml
 .fake
 .vs
 .vscode
-*.DotSettings.user
