+++ Domain.Test/BacklogStateTest.cs
@@ -1,5 +1,4 @@
 using Domain.Exceptions;
-using Domain.Sprints;
 using NSubstitute;
 
 namespace Domain.Test;
@@ -11,14 +10,11 @@ public class BacklogStateTest
     {
         var writer = Substitute.For<IWriter>();
         var notificationWriter = Substitute.For<IWriter>();
-        
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
-        project.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
+        sprint.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
        
         backlogItem.ToDoing();
         backlogItem.ToReadyForTesting();
@@ -36,14 +32,11 @@ public class BacklogStateTest
         var writer = Substitute.For<IWriter>();
         var notificationWriter = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
-            
-        project.Tester.Subscribe(new EmailNotifier("henkdetesterman@gmail.com", notificationWriter));
+        
+        sprint.Tester.Subscribe(new EmailNotifier("henkdetesterman@gmail.com", notificationWriter));
         
         backlogItem.ToDoing();
         backlogItem.ToReadyForTesting();
@@ -57,14 +50,11 @@ public class BacklogStateTest
         var writer = Substitute.For<IWriter>();
         var notificationWriter = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
         
-        project.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
+        sprint.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
         
         backlogItem.ToDoing();
         backlogItem.ToReadyForTesting();
@@ -78,14 +68,11 @@ public class BacklogStateTest
         var writer = Substitute.For<IWriter>();
         var notificationWriter = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
         
-        project.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
+        sprint.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
         
         backlogItem.ToDoing();
 
@@ -98,14 +85,11 @@ public class BacklogStateTest
         var writer = Substitute.For<IWriter>();
         var notificationWriter = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
         
-        project.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
+        sprint.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
         
         backlogItem.ToDoing();
         backlogItem.ToReadyForTesting();
@@ -120,14 +104,11 @@ public class BacklogStateTest
         var writer = Substitute.For<IWriter>();
         var notificationWriter = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
         
-        project.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
+        sprint.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", notificationWriter));
         
         IllegalStateAdvanceException ex = Assert.Throws<IllegalStateAdvanceException>(
             () => backlogItem.ToTodo());
@@ -139,11 +120,8 @@ public class BacklogStateTest
     {
         var writer = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var backlogItem = new BacklogItem("1", writer, sprint, new TeamMember("Linus Torvalds"));
         var activity = new BacklogItem("2", writer, sprint, new TeamMember("Henk de steen"));
         backlogItem.Activities.Add(activity);
