+++ Domain.Test/TeamMemberNotifierTest.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using Domain.Sprints;
 using NSubstitute;
 
 namespace Domain.Test;
@@ -11,13 +10,10 @@ public class TeamMemberNotifierTest
     {
         var writer = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
-        project.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", writer));
-        project.ScrumMaster.Notify("Hello scrummaster!");
+        sprint.ScrumMaster.Subscribe(new EmailNotifier("jandescrumman@gmail.com", writer));
+        sprint.ScrumMaster.Notify("Hello scrummaster!");
 
         writer.Received().WriteLine("To: Jan de Scrumman <jandescrumman@gmail.com>: Hello scrummaster!");
     }
@@ -26,16 +22,13 @@ public class TeamMemberNotifierTest
     {
         var writer = Substitute.For<IWriter>();
 
-        var project = new Project("SO&A 2",new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
+        var sprint = new Sprint(new TeamMember("Jan de Scrumman"), new TeamMember("Henk de Testerman"),
             new TeamMember("Jan de Productowner"));
-        var sprintFactory = new SprintFactory();
-        var sprint = sprintFactory.NewReleaseSprint(project);
-        
         var emailNotifier = new EmailNotifier("jandescrumman@gmail.com", writer);
-        var unsubscriber = project.ScrumMaster.Subscribe(emailNotifier);
+        var unsubscriber = sprint.ScrumMaster.Subscribe(emailNotifier);
         unsubscriber.Dispose();
         
-        project.ScrumMaster.Notify("Hello scrummaster!");
+        sprint.ScrumMaster.Notify("Hello scrummaster!");
         
         writer.DidNotReceive().WriteLine("To: Jan de Scrumman <jandescrumman@gmail.com>: Hello scrummaster!");
     }
